requirements.txt
alembic==1.8.1
anyio==3.6.1
certifi==2022.6.15
chardet==3.0.4
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
fastapi==0.79.0
googletrans==4.0.0rc1
greenlet==1.1.3
gunicorn==20.1.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.7.10
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Mako==1.2.1
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2==2.9.3
pyasn1==0.4.8
pydantic==1.10.0a1
python-dotenv==0.20.0
python-jose==3.3.0
python-multipart==0.0.5
requests==2.28.1
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.3.0
urllib3==1.26.10
uvicorn==0.18.2
